backend/requirements/base.txt
django==4.2.11
django-environ==0.11.2
djangorestframework==3.15.1
django-cors-headers==4.3.1 
drf-spectacular==0.27.2  
argon2-cffi==23.1.0
pillow==10.3.0